Move PaymentSendFailure into outbound_payment module
And re-export it in channelmanager.rs so it can remain public
